test: strengthen unified compose coverage
