+++ apps/web/__tests__/playwright/emulated/mail/local-cache.spec.ts
@@ -1,6 +1,8 @@
-import { randomUUID } from "node:crypto";
 import { expect, test, type Page } from "@playwright/test";
-import { Client } from "pg";
+import {
+  createSecondEmailAccount,
+  deleteSecondEmailAccount,
+} from "./account-test-helpers";
 import { conversationWithSubject, openMail } from "./mail-test-helpers";
 
 const MESSAGE_COUNT = 5000;
@@ -507,54 +509,4 @@ function readerBody(page: Page, text: string) {
   return page
     .locator("pre")
     .filter({ hasText: new RegExp(`^${escapeRegExp(text)}$`) });
-}
-
-async function createSecondEmailAccount(primaryEmailAccountId: string) {
-  const client = new Client({ connectionString: process.env.DATABASE_URL });
-  await client.connect();
-  const suffix = randomUUID();
-  const accountId = `playwright-account-${suffix}`;
-  const id = `playwright-email-account-${suffix}`;
-  const email = `playwright-secondary-${suffix}@gmail.com`;
-  const name = "Playwright Secondary";
-
-  try {
-    await client.query("BEGIN");
-    const userResult = await client.query<{ userId: string }>(
-      `SELECT "userId" FROM "EmailAccount" WHERE id = $1`,
-      [primaryEmailAccountId],
-    );
-    const userId = userResult.rows[0]?.userId;
-    if (!userId) throw new Error("Could not find the Playwright account user");
-
-    await client.query(
-      `INSERT INTO "Account"
-        (id, "createdAt", "updatedAt", "userId", provider, type, "providerAccountId")
-       VALUES ($1, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP, $2, 'google', 'oidc', $3)`,
-      [accountId, userId, `playwright-provider-${suffix}`],
-    );
-    await client.query(
-      `INSERT INTO "EmailAccount"
-        (id, email, name, "createdAt", "updatedAt", "userId", "accountId")
-       VALUES ($1, $2, $3, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP, $4, $5)`,
-      [id, email, name, userId, accountId],
-    );
-    await client.query("COMMIT");
-    return { accountId, email, id, name };
-  } catch (error) {
-    await client.query("ROLLBACK");
-    throw error;
-  } finally {
-    await client.end();
-  }
-}
-
-async function deleteSecondEmailAccount(accountId: string) {
-  const client = new Client({ connectionString: process.env.DATABASE_URL });
-  await client.connect();
-  try {
-    await client.query(`DELETE FROM "Account" WHERE id = $1`, [accountId]);
-  } finally {
-    await client.end();
-  }
 }
